requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
python-dotenv==1.1.1
rapidfuzz==3.14.1
ftfy==6.3.1
aiohttp==3.12.15
python-multipart==0.0.20
pandas==2.2.3
numpy>=1.24.0
scikit-learn==1.5.2
pyarrow==18.0.0
